+++ src/main/java/com/springbootproject/service/TeacherService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface TeacherService {
     List<Teacher> getAllTeachers();
     Teacher saveTeacher(TeacherDto teacherDto);
-    Teacher getTeacherById(Long id);
+    Teacher getTeacherById(int id);
     Teacher updateTeacher(Teacher teacher);
-    void deleteTeacherById(Long id);
+    void deleteTeacherById(int id);
 }
